feat: refactor backend architecture with interface-based design
- Add backend interface definitions in backend-interface.lua
- Extract snacks backend implementation to separate snacks-backend.lua
- Rename tmux.lua to tmux-backend.lua for consistency
- Update terminal.lua to use unified backend delegation pattern
- Replace direct method calls with object-oriented terminal API
- Implement TerminalObject base class with backend-specific subclasses
- Standardize error handling and session management across backends
